requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
Faker==19.10.0
Markdown==3.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
validate-docbr==1.10.0
